+++ packages/react/src/exports/connectors.test.ts
@@ -7,7 +7,6 @@ test('exports', () => {
     [
       "baseAccount",
       "coinbaseWallet",
-      "gemini",
       "injected",
       "metaMask",
       "mock",