+++ packages/react/src/tempo/hooks/fee.test.ts
@@ -12,7 +12,7 @@ describe('useUserToken', () => {
       { initialProps: { account: undefined as Address | undefined } },
     )
 
-    await vi.waitFor(() => result.current.fetchStatus === 'fetching')
+    await vi.waitFor(() => expect(result.current.fetchStatus).toBe('idle'))
 
     expect(result.current).toMatchInlineSnapshot(`
       {
@@ -53,7 +53,7 @@ describe('useUserToken', () => {
     rerender({ account: accounts[0].address })
 
     await vi.waitFor(() => expect(result.current.isSuccess).toBeTruthy(), {
-      timeout: 5_000,
+      timeout: 10_000,
     })
 
     expect(result.current).toMatchInlineSnapshot(`